+++ src/storage-helper.js
@@ -8,11 +8,13 @@ module.exports = (function storageHelper() {
             return new Promise(resolve => {
                 chrome.storage.sync.get({
                     highlightSyntax: true,
-                    highlightOcurrences: true
+                    highlightOcurrences: true,
+                    autocollapsePaths: ['package-lock.json']
                 }, options => {
                     resolve({
                         highlightSyntax: options.highlightSyntax,
-                        highlightOcurrences: options.highlightOcurrences
+                        highlightOcurrences: options.highlightOcurrences,
+                        autocollapsePaths: options.autocollapsePaths
                     });
                 });
             });